chore: move theme.js into frontend/theme.js and update import
