+++ pages/_app.js
@@ -1,7 +1,7 @@
 import * as React from "react"
 import { ChakraProvider } from "@chakra-ui/react"
 
-import theme from "theme"
+import theme from "frontend/theme"
 
 function App({ Component, pageProps }) {
   return (
